requirements.txt
charset-normalizer==3.1.0
elementpath==4.0.1
grobid-client-python==0.0.5
idna==3.4
pillow==9.4.0
pyparsing==3.0.9
python-dateutil==2.8.2
requests==2.28.2
six==1.16.0
urllib3==1.26.14
wordcloud==1.8.2.2
